Move etcd tools to components (#688)
